Merge changes from 2.3.0 to 2.3.1
The development branch was on MOODLE_STABLE_2.3 just after v2.3.0.
When rebasing the branch, some conflicts arose which were postponed
to this commit.

The main change is that the definition of a field in a form can now
eventually receive an "objectid" param. The user extension uses this
parameter to determine is the "required" state must be enforced.